+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Version 1.0.0
 - Turn :meth:`Element.unserialize` into an instance method.
 - Rename :meth:`Element.set` to `Element.set_from_raw`.
 - Add :meth:`Element.serialize` and :meth:`Element.set_from_native`.
+- Make :data:`Unspecified` and :data:`NotUnserializable` false-ish.
 
 Version 0.2.0
 -------------